change the style of the customized button
